feat:Division de los steps en CardScheduleAppointmentPrimary y diseño responsivo en el mismo
